fix poll, and other issues

+++ src/lib/showpollitem.svelte
@@ -5,6 +5,7 @@
   import { slide } from "svelte/transition";
   import Editpoll from "./editpoll.svelte";
   import ShowPollOption from "./showpolloption.svelte";
+  import { current_group } from "./stores/groups";
 
   export let poll: Poll;
   let showedit: boolean = false;
@@ -43,24 +44,27 @@
         {poll.total_vote} total votes
       </p>
     </div>
-    <div class="flex justify-end items-center col-span-3 h-full">
-      <button
-        type="button"
-        class="text-white bg-blue-700 dark:bg-blue-600 hover:bg-blue-800 dark:hover:bg-blue-700 focus:ring-4 focus:ring-blue-300 font-medium transition-shadow rounded-lg text-sm px-5 py-3 text-center w-1/10 mx-2"
-        on:click={() => {
-          showedit = true;
-        }}
-      >
-        <FaIcon icon="pencil" />&nbsp;&nbsp;Edit Poll
-      </button>
-      <button
-        type="button"
-        class="text-white bg-red-700 dark:bg-red-600 hover:bg-red-800 dark:hover:bg-red-700 focus:ring-4 focus:ring-red-300 font-medium transition-shadow rounded-lg text-sm px-5 py-3 text-center w-1/10 mx-2"
-        on:click={() => deletePoll(poll.id)}
-      >
-        <FaIcon icon="trash" />&nbsp;&nbsp;Delete Poll
-      </button>
-    </div>
+
+    {#if $current_group && $current_group.access === "admin"}
+      <div class="flex justify-end items-center col-span-3 h-full">
+        <button
+          type="button"
+          class="text-white bg-blue-700 dark:bg-blue-600 hover:bg-blue-800 dark:hover:bg-blue-700 focus:ring-4 focus:ring-blue-300 font-medium transition-shadow rounded-lg text-sm px-5 py-3 text-center w-1/10 mx-2"
+          on:click={() => {
+            showedit = true;
+          }}
+        >
+          <FaIcon icon="pencil" />&nbsp;&nbsp;Edit Poll
+        </button>
+        <button
+          type="button"
+          class="text-white bg-red-700 dark:bg-red-600 hover:bg-red-800 dark:hover:bg-red-700 focus:ring-4 focus:ring-red-300 font-medium transition-shadow rounded-lg text-sm px-5 py-3 text-center w-1/10 mx-2"
+          on:click={() => deletePoll(poll.id)}
+        >
+          <FaIcon icon="trash" />&nbsp;&nbsp;Delete Poll
+        </button>
+      </div>
+    {/if}
   </div>
   <!-- if you use auto-rows-fr, then opening one option will cause all others to expand (but not showing option), because all rows have to be of same width -->
   <!-- but then again, if you do not use it, then voted or unvoted options will be of different widths, leading to mismatched version -->
